refactor(monitor): extract shutdown + health phases into mixins
Split UPSGroupMonitor's god-object body across two new subpackages so
each phase is independently readable:

* eneru.shutdown.vms.VMShutdownMixin           - libvirt VMs
* eneru.shutdown.containers.ContainerShutdownMixin - docker/podman + compose
* eneru.shutdown.filesystems.FilesystemShutdownMixin - sync + unmount
* eneru.shutdown.remote.RemoteShutdownMixin    - SSH-based remote servers
* eneru.health.voltage.VoltageMonitorMixin     - thresholds, AVR, bypass, overload
* eneru.health.battery.BatteryMonitorMixin     - depletion rate, anomaly detection

UPSGroupMonitor inherits all six mixins; method bodies are moved
verbatim. monitor.py shrinks from 1,813 to ~830 lines and keeps only
the core: init/dependency checks, UPS polling, save/log/notify,
shutdown orchestration (_execute_shutdown_sequence), main loop, and
on-battery / on-line / connection-failure handlers.

Stale imports pruned now that monitor.py no longer needs
REMOTE_ACTIONS, deque, the dataclass type names, TimezoneFormatter,
ConfigLoader, or YAML_AVAILABLE.

nfpm.yaml updated with explicit contents entries for all 8 new module
files plus the two subpackage directories. Verified by simulating the
/opt/ups-monitor layout with an isolated interpreter (caught the same
class of bug as PR #23).

Test patches updated where they targeted eneru.monitor.run_command for
methods that have moved into mixins (TestRunRemoteCommand class +
TestShutdownSequence.test_shutdown_flag_file_created_during_sequence).
