+++ src/eneru/monitor.py
@@ -13,35 +13,19 @@ import threading
 from datetime import datetime
 from pathlib import Path
 from typing import Optional, Dict, Tuple, List
-from collections import deque
 
-# Import from new modules
 from eneru.version import __version__
-from eneru.config import (
-    Config,
-    UPSConfig,
-    TriggersConfig,
-    DepletionConfig,
-    ExtendedTimeConfig,
-    BehaviorConfig,
-    LoggingConfig,
-    NotificationsConfig,
-    VMConfig,
-    ContainersConfig,
-    ComposeFileConfig,
-    FilesystemsConfig,
-    UnmountConfig,
-    RemoteServerConfig,
-    RemoteCommandConfig,
-    LocalShutdownConfig,
-    ConfigLoader,
-    YAML_AVAILABLE,
-)
+from eneru.config import Config, RemoteServerConfig
 from eneru.state import MonitorState
-from eneru.logger import UPSLogger, TimezoneFormatter
+from eneru.logger import UPSLogger
 from eneru.notifications import NotificationWorker, APPRISE_AVAILABLE
 from eneru.utils import run_command, command_exists, is_numeric, format_seconds
-from eneru.actions import REMOTE_ACTIONS
+from eneru.shutdown.vms import VMShutdownMixin
+from eneru.shutdown.containers import ContainerShutdownMixin
+from eneru.shutdown.filesystems import FilesystemShutdownMixin
+from eneru.shutdown.remote import RemoteShutdownMixin
+from eneru.health.voltage import VoltageMonitorMixin
+from eneru.health.battery import BatteryMonitorMixin
 
 # Re-export for backwards compatibility (tests may mock these)
 try:
@@ -86,7 +70,14 @@ def compute_effective_order(
     return result
 
 
-class UPSGroupMonitor:
+class UPSGroupMonitor(
+    VMShutdownMixin,
+    ContainerShutdownMixin,
+    FilesystemShutdownMixin,
+    RemoteShutdownMixin,
+    VoltageMonitorMixin,
+    BatteryMonitorMixin,
+):
     """Main UPS Monitor class."""
 
     def __init__(self, config: Config, exit_after_shutdown: bool = False,
@@ -415,45 +406,6 @@ class UPSGroupMonitor:
             for server in self.config.remote_servers:
                 server.enabled = False
 
-    def _detect_container_runtime(self) -> Optional[str]:
-        """Detect available container runtime."""
-        runtime_config = self.config.containers.runtime.lower()
-
-        if runtime_config == "docker":
-            if command_exists("docker"):
-                return "docker"
-            self._log_message("⚠️ WARNING: Docker specified but not found")
-            return None
-
-        elif runtime_config == "podman":
-            if command_exists("podman"):
-                return "podman"
-            self._log_message("⚠️ WARNING: Podman specified but not found")
-            return None
-
-        elif runtime_config == "auto":
-            if command_exists("podman"):
-                return "podman"
-            elif command_exists("docker"):
-                return "docker"
-            return None
-
-        else:
-            self._log_message(f"⚠️ WARNING: Unknown container runtime '{runtime_config}'")
-            return None
-
-    def _check_compose_available(self) -> bool:
-        """Check if compose subcommand is available for the detected runtime."""
-        if not self._container_runtime:
-            return False
-
-        # Try running 'docker/podman compose version' to check availability
-        exit_code, _, _ = run_command(
-            [self._container_runtime, "compose", "version"],
-            timeout=10
-        )
-        return exit_code == 0
-
     def _get_ups_var(self, var_name: str) -> Optional[str]:
         """Get a single UPS variable using upsc."""
         exit_code, stdout, _ = run_command(["upsc", self.config.ups.name, var_name])
@@ -479,33 +431,6 @@ class UPSGroupMonitor:
 
         return True, ups_data, ""
 
-    def _initialize_voltage_thresholds(self):
-        """Initialize voltage thresholds dynamically from UPS data."""
-        nominal = self._get_ups_var("input.voltage.nominal")
-        low_transfer = self._get_ups_var("input.transfer.low")
-        high_transfer = self._get_ups_var("input.transfer.high")
-
-        if is_numeric(nominal):
-            self.state.nominal_voltage = float(nominal)
-        else:
-            self.state.nominal_voltage = 230.0
-
-        if is_numeric(low_transfer):
-            self.state.voltage_warning_low = float(low_transfer) + 5
-        else:
-            self.state.voltage_warning_low = self.state.nominal_voltage * 0.9
-
-        if is_numeric(high_transfer):
-            self.state.voltage_warning_high = float(high_transfer) - 5
-        else:
-            self.state.voltage_warning_high = self.state.nominal_voltage * 1.1
-
-        self._log_message(
-            f"📊 Voltage Monitoring Active. Nominal: {self.state.nominal_voltage}V. "
-            f"Low Warning: {self.state.voltage_warning_low}V. "
-            f"High Warning: {self.state.voltage_warning_high}V."
-        )
-
     def _wait_for_initial_connection(self):
         """Wait for initial connection to NUT server."""
         self._log_message(f"⏳ Checking initial connection to {self.config.ups.name}...")
@@ -530,44 +455,6 @@ class UPSGroupMonitor:
                 f"within {max_wait}s. Proceeding, but voltage thresholds may default."
             )
 
-    def _calculate_depletion_rate(self, current_battery: str) -> float:
-        """Calculate battery depletion rate based on history."""
-        current_time = int(time.time())
-
-        if not is_numeric(current_battery):
-            return 0.0
-
-        current_battery_float = float(current_battery)
-        cutoff_time = current_time - self.config.triggers.depletion.window
-
-        self.state.battery_history = deque(
-            [(ts, bat) for ts, bat in self.state.battery_history if ts >= cutoff_time],
-            maxlen=1000
-        )
-        self.state.battery_history.append((current_time, current_battery_float))
-
-        try:
-            temp_file = self._battery_history_path.with_suffix('.tmp')
-            with open(temp_file, 'w') as f:
-                for ts, bat in self.state.battery_history:
-                    f.write(f"{ts}:{bat}\n")
-            temp_file.replace(self._battery_history_path)
-        except Exception:
-            pass
-
-        if len(self.state.battery_history) < 30:
-            return 0.0
-
-        oldest_time, oldest_battery = self.state.battery_history[0]
-        time_diff = current_time - oldest_time
-
-        if time_diff > 0:
-            battery_diff = oldest_battery - current_battery_float
-            rate = (battery_diff / time_diff) * 60
-            return round(rate, 2)
-
-        return 0.0
-
     def _save_state(self, ups_data: Dict[str, str]):
         """Save current UPS state to file."""
         state_content = (
@@ -585,592 +472,6 @@ class UPSGroupMonitor:
             temp_file.replace(self._state_file_path)
         except Exception:
             pass
-
-    # ==========================================================================
-    # SHUTDOWN SEQUENCE
-    # ==========================================================================
-
-    def _shutdown_vms(self):
-        """Shutdown all libvirt virtual machines."""
-        if not self.config.virtual_machines.enabled:
-            return
-
-        self._log_message("🖥️ Shutting down all libvirt virtual machines...")
-
-        if not command_exists("virsh"):
-            self._log_message("  ℹ️ virsh not available, skipping VM shutdown")
-            return
-
-        exit_code, stdout, _ = run_command(["virsh", "list", "--name", "--state-running"])
-        if exit_code != 0:
-            self._log_message("  ⚠️ Failed to get VM list")
-            return
-
-        running_vms = [vm.strip() for vm in stdout.strip().split('\n') if vm.strip()]
-
-        if not running_vms:
-            self._log_message("  ℹ️ No running VMs found")
-            return
-
-        for vm in running_vms:
-            self._log_message(f"  ⏹️ Shutting down VM: {vm}")
-            if self.config.behavior.dry_run:
-                self._log_message(f"  🧪 [DRY-RUN] Would shutdown VM: {vm}")
-            else:
-                exit_code, stdout, stderr = run_command(["virsh", "shutdown", vm])
-                if stdout.strip():
-                    self._log_message(f"    {stdout.strip()}")
-
-        if self.config.behavior.dry_run:
-            return
-
-        max_wait = self.config.virtual_machines.max_wait
-        self._log_message(f"  ⏳ Waiting up to {max_wait}s for VMs to shutdown gracefully...")
-        wait_interval = 5
-        time_waited = 0
-        remaining_vms: List[str] = []
-
-        while time_waited < max_wait:
-            exit_code, stdout, _ = run_command(["virsh", "list", "--name", "--state-running"])
-            still_running = set(vm.strip() for vm in stdout.strip().split('\n') if vm.strip())
-            remaining_vms = [vm for vm in running_vms if vm in still_running]
-
-            if not remaining_vms:
-                self._log_message(f"  ✅ All VMs stopped gracefully after {time_waited}s.")
-                break
-
-            self._log_message(f"  🕒 Still waiting for: {' '.join(remaining_vms)} (Waited {time_waited}s)")
-            time.sleep(wait_interval)
-            time_waited += wait_interval
-
-        if remaining_vms:
-            self._log_message("  ⚠️ Timeout reached. Force destroying remaining VMs.")
-            for vm in remaining_vms:
-                self._log_message(f"  ⚡ Force destroying VM: {vm}")
-                run_command(["virsh", "destroy", vm])
-
-        self._log_message("  ✅ All VMs shutdown complete")
-
-    def _shutdown_compose_stacks(self):
-        """Shutdown docker/podman compose stacks in order (best effort)."""
-        if not self._compose_available:
-            return
-
-        if not self.config.containers.compose_files:
-            return
-
-        runtime = self._container_runtime
-        runtime_display = runtime.capitalize()
-
-        self._log_message(
-            f"🐳 Stopping {runtime_display} Compose stacks "
-            f"({len(self.config.containers.compose_files)} file(s))..."
-        )
-
-        for compose_file in self.config.containers.compose_files:
-            file_path = compose_file.path
-            if not file_path:
-                continue
-
-            # Determine timeout: per-file or global
-            timeout = compose_file.stop_timeout
-            if timeout is None:
-                timeout = self.config.containers.stop_timeout
-
-            # Check if file exists (best effort - warn if not)
-            if not Path(file_path).exists():
-                self._log_message(f"  ⚠️ Compose file not found: {file_path} (skipping)")
-                continue
-
-            self._log_message(f"  ➡️ Stopping: {file_path} (timeout: {timeout}s)")
-
-            if self.config.behavior.dry_run:
-                self._log_message(
-                    f"  🧪 [DRY-RUN] Would execute: {runtime} compose -f {file_path} down"
-                )
-                continue
-
-            # Run compose down
-            compose_cmd = [runtime, "compose", "-f", file_path, "down"]
-            exit_code, stdout, stderr = run_command(compose_cmd, timeout=timeout + 30)
-
-            if exit_code == 0:
-                self._log_message(f"  ✅ {file_path} stopped successfully")
-            elif exit_code == 124:
-                self._log_message(
-                    f"  ⚠️ {file_path} compose down timed out after {timeout}s (continuing)"
-                )
-            else:
-                error_msg = stderr.strip() if stderr.strip() else f"exit code {exit_code}"
-                self._log_message(f"  ⚠️ {file_path} compose down failed: {error_msg} (continuing)")
-
-        self._log_message("  ✅ Compose stacks shutdown complete")
-
-    def _shutdown_containers(self):
-        """Stop all containers using detected runtime (Docker/Podman).
-
-        Execution order:
-        1. Shutdown compose stacks (best effort, in order)
-        2. Shutdown all remaining containers (if shutdown_all_remaining_containers is True)
-        """
-        if not self.config.containers.enabled:
-            return
-
-        if not self._container_runtime:
-            return
-
-        runtime = self._container_runtime
-        runtime_display = runtime.capitalize()
-
-        # Phase 1: Shutdown compose stacks first (best effort)
-        self._shutdown_compose_stacks()
-
-        # Phase 2: Shutdown all remaining containers
-        if not self.config.containers.shutdown_all_remaining_containers:
-            self._log_message(f"🐳 Skipping remaining {runtime_display} container shutdown (disabled)")
-            return
-
-        self._log_message(f"🐳 Stopping all remaining {runtime_display} containers...")
-
-        # Get list of running containers
-        exit_code, stdout, _ = run_command([runtime, "ps", "-q"])
-        if exit_code != 0:
-            self._log_message(f"  ⚠️ Failed to get {runtime_display} container list")
-            return
-
-        container_ids = [cid.strip() for cid in stdout.strip().split('\n') if cid.strip()]
-
-        if not container_ids:
-            self._log_message(f"  ℹ️ No running {runtime_display} containers found")
-        else:
-            if self.config.behavior.dry_run:
-                exit_code, stdout, _ = run_command([runtime, "ps", "--format", "{{.Names}}"])
-                names = stdout.strip().replace('\n', ' ')
-                self._log_message(f"  🧪 [DRY-RUN] Would stop {runtime_display} containers: {names}")
-            else:
-                # Stop containers with timeout
-                stop_cmd = [runtime, "stop", "-t", str(self.config.containers.stop_timeout)]
-                stop_cmd.extend(container_ids)
-                run_command(stop_cmd, timeout=self.config.containers.stop_timeout + 30)
-                self._log_message(f"  ✅ {runtime_display} containers stopped")
-
-        # Handle Podman rootless containers if configured
-        if runtime == "podman" and self.config.containers.include_user_containers:
-            self._shutdown_podman_user_containers()
-
-    def _shutdown_podman_user_containers(self):
-        """Stop Podman containers running as non-root users."""
-        self._log_message("  🔍 Checking for rootless Podman containers...")
-
-        if self.config.behavior.dry_run:
-            self._log_message("  🧪 [DRY-RUN] Would stop rootless Podman containers for all users")
-            return
-
-        # Get list of users with active Podman containers
-        # This requires loginctl and users with linger enabled
-        exit_code, stdout, _ = run_command(["loginctl", "list-users", "--no-legend"])
-        if exit_code != 0:
-            self._log_message("  ⚠️ Failed to list users for rootless container check")
-            return
-
-        for line in stdout.strip().split('\n'):
-            if not line.strip():
-                continue
-
-            parts = line.split()
-            if len(parts) >= 2:
-                uid = parts[0]
-                username = parts[1]
-
-                # Skip system users (UID < 1000)
-                try:
-                    if int(uid) < 1000:
-                        continue
-                except ValueError:
-                    continue
-
-                # Check for running containers as this user
-                exit_code, stdout, _ = run_command([
-                    "sudo", "-u", username,
-                    "podman", "ps", "-q"
-                ], timeout=10)
-
-                if exit_code == 0 and stdout.strip():
-                    container_ids = [cid.strip() for cid in stdout.strip().split('\n') if cid.strip()]
-                    if container_ids:
-                        self._log_message(f"  👤 Stopping {len(container_ids)} container(s) for user '{username}'")
-                        stop_cmd = [
-                            "sudo", "-u", username,
-                            "podman", "stop", "-t", str(self.config.containers.stop_timeout)
-                        ]
-                        stop_cmd.extend(container_ids)
-                        run_command(stop_cmd, timeout=self.config.containers.stop_timeout + 30)
-
-        self._log_message("  ✅ Rootless Podman containers stopped")
-
-    def _sync_filesystems(self):
-        """Sync all filesystems.
-
-        Note: os.sync() schedules buffers to be flushed but may return before
-        physical write completion on some systems. The 2-second sleep allows
-        storage controllers (especially battery-backed RAID) to flush their
-        write-back caches before power is cut.
-        """
-        if not self.config.filesystems.sync_enabled:
-            return
-
-        self._log_message("💾 Syncing all filesystems...")
-        if self.config.behavior.dry_run:
-            self._log_message("  🧪 [DRY-RUN] Would sync filesystems")
-        else:
-            os.sync()
-            time.sleep(2)  # Allow storage controller caches to flush
-            self._log_message("  ✅ Filesystems synced")
-
-    def _unmount_filesystems(self):
-        """Unmount configured filesystems."""
-        if not self.config.filesystems.unmount.enabled:
-            return
-
-        if not self.config.filesystems.unmount.mounts:
-            return
-
-        timeout = self.config.filesystems.unmount.timeout
-        self._log_message(f"📤 Unmounting filesystems (Max wait: {timeout}s)...")
-
-        for mount in self.config.filesystems.unmount.mounts:
-            mount_point = mount.get('path', '')
-            options = mount.get('options', '')
-
-            if not mount_point:
-                continue
-
-            options_display = f" {options}" if options else ""
-            self._log_message(f"  ➡️ Unmounting {mount_point}{options_display}")
-
-            if self.config.behavior.dry_run:
-                self._log_message(
-                    f"  🧪 [DRY-RUN] Would execute: timeout {timeout}s umount {options} {mount_point}"
-                )
-                continue
-
-            cmd = ["umount"]
-            if options:
-                cmd.append(options)
-            cmd.append(mount_point)
-
-            exit_code, _, stderr = run_command(cmd, timeout=timeout)
-
-            if exit_code == 0:
-                self._log_message(f"  ✅ {mount_point} unmounted successfully")
-            elif exit_code == 124:
-                self._log_message(
-                    f"  ⚠️ {mount_point} unmount timed out "
-                    "(device may be busy/unreachable). Proceeding anyway."
-                )
-            else:
-                check_code, _, _ = run_command(["mountpoint", "-q", mount_point])
-                if check_code == 0:
-                    self._log_message(
-                        f"  ❌ Failed to unmount {mount_point} "
-                        f"(Error code {exit_code}). Proceeding anyway."
-                    )
-                else:
-                    self._log_message(f"  ℹ️ {mount_point} was likely not mounted.")
-
-    def _shutdown_remote_servers(self):
-        """Shutdown all enabled remote servers via SSH.
-
-        Servers are grouped by their effective shutdown order and processed
-        in ascending order.  All servers within a group run in parallel.
-        A server alone in its group effectively runs sequentially.
-
-        When shutdown_order is not set, the legacy parallel flag determines
-        effective order:
-        - parallel: true  (default) -> effective order 0
-        - parallel: false -> unique negative orders (run before order 0)
-        This preserves exact backward compatibility with existing configs.
-        """
-        enabled_servers = [s for s in self.config.remote_servers if s.enabled]
-
-        if not enabled_servers:
-            return
-
-        # Group servers by effective shutdown order
-        ordered = compute_effective_order(enabled_servers)
-        phases: Dict[int, List[RemoteServerConfig]] = {}
-        for effective, server in ordered:
-            phases.setdefault(effective, []).append(server)
-        sorted_keys = sorted(phases.keys())
-
-        server_count = len(enabled_servers)
-        num_phases = len(sorted_keys)
-
-        if num_phases > 1:
-            self._log_message(
-                f"🌐 Shutting down {server_count} remote server(s) in {num_phases} phases..."
-            )
-        elif server_count > 1:
-            self._log_message(f"🌐 Shutting down {server_count} remote server(s) in parallel...")
-        else:
-            self._log_message(f"🌐 Shutting down 1 remote server...")
-
-        completed = 0
-
-        for phase_idx, key in enumerate(sorted_keys, 1):
-            phase_servers = phases[key]
-            names = ", ".join(s.name or s.host for s in phase_servers)
-
-            if num_phases > 1:
-                self._log_message(f"  📋 Phase {phase_idx}/{num_phases} (order={key}): {names}")
-
-            if len(phase_servers) == 1:
-                server = phase_servers[0]
-                display_name = server.name or server.host
-                try:
-                    self._shutdown_remote_server(server)
-                    completed += 1
-                except Exception as e:
-                    self._log_message(f"  ❌ {display_name} shutdown failed: {e}")
-            else:
-                completed += self._shutdown_servers_parallel(phase_servers)
-
-        # Log summary
-        self._log_message(f"  ✅ Remote shutdown complete ({completed}/{server_count} servers)")
-
-    def _shutdown_servers_parallel(self, servers: List[RemoteServerConfig]) -> int:
-        """Shutdown multiple remote servers in parallel using threads.
-
-        Returns the number of servers whose threads finished within the
-        timeout window (regardless of individual success/failure — per-server
-        errors are logged inside _shutdown_remote_server).
-        """
-        def calc_server_timeout(server: RemoteServerConfig) -> int:
-            pre_cmd_time = sum(
-                (cmd.timeout or server.command_timeout) for cmd in server.pre_shutdown_commands
-            )
-            return (
-                pre_cmd_time
-                + server.command_timeout
-                + server.connect_timeout
-                + server.shutdown_safety_margin
-            )
-
-        max_timeout = max(calc_server_timeout(s) for s in servers)
-
-        def shutdown_server_thread(server: RemoteServerConfig):
-            """Thread worker for shutting down a single server."""
-            try:
-                self._shutdown_remote_server(server)
-            except Exception:
-                pass  # Errors logged inside _shutdown_remote_server
-
-        threads: List[threading.Thread] = []
-        for server in servers:
-            t = threading.Thread(
-                target=shutdown_server_thread,
-                args=(server,),
-                name=f"remote-shutdown-{server.name or server.host}"
-            )
-            t.start()
-            threads.append(t)
-
-        # Deadline-based join: cap total wait at max_timeout regardless of
-        # how many threads are stuck. Per-thread join() with the same
-        # max_timeout would stack to N × max_timeout in the worst case.
-        deadline = time.monotonic() + max_timeout
-        for t in threads:
-            remaining = max(0.0, deadline - time.monotonic())
-            t.join(timeout=remaining)
-
-        still_running = [t for t in threads if t.is_alive()]
-        if still_running:
-            self._log_message(
-                f"  ⚠️ {len(still_running)} remote shutdown(s) still in progress "
-                "(continuing with next phase)"
-            )
-
-        return len(servers) - len(still_running)
-
-    def _run_remote_command(
-        self,
-        server: RemoteServerConfig,
-        command: str,
-        timeout: int,
-        description: str
-    ) -> Tuple[bool, str]:
-        """Run a single command on a remote server via SSH.
-
-        Returns:
-            Tuple of (success, error_message)
-        """
-        display_name = server.name or server.host
-
-        ssh_cmd = ["ssh"]
-
-        # Add configured SSH options
-        for opt in server.ssh_options:
-            if opt.startswith("-o"):
-                ssh_cmd.append(opt)
-            else:
-                ssh_cmd.extend(["-o", opt])
-
-        ssh_cmd.extend([
-            "-o", f"ConnectTimeout={server.connect_timeout}",
-            "-o", "BatchMode=yes",  # Prevent password prompts from hanging
-            f"{server.user}@{server.host}",
-            command
-        ])
-
-        # Add buffer to timeout to account for SSH connection overhead
-        exit_code, stdout, stderr = run_command(ssh_cmd, timeout=timeout + 30)
-
-        if exit_code == 0:
-            return True, ""
-        elif exit_code == 124:
-            return False, f"timed out after {timeout}s"
-        else:
-            error_msg = stderr.strip() if stderr.strip() else f"exit code {exit_code}"
-            return False, error_msg
-
-    def _execute_remote_pre_shutdown(self, server: RemoteServerConfig) -> bool:
-        """Execute pre-shutdown commands on a remote server.
-
-        Returns:
-            True if all commands executed (success or best-effort failure)
-            False if SSH connection failed entirely
-        """
-        if not server.pre_shutdown_commands:
-            return True
-
-        display_name = server.name or server.host
-        cmd_count = len(server.pre_shutdown_commands)
-
-        self._log_message(f"  📋 Executing {cmd_count} pre-shutdown command(s)...")
-
-        for idx, cmd_config in enumerate(server.pre_shutdown_commands, 1):
-            # Determine timeout
-            timeout = cmd_config.timeout
-            if timeout is None:
-                timeout = server.command_timeout
-
-            # Handle predefined action
-            if cmd_config.action:
-                action_name = cmd_config.action.lower()
-
-                if action_name not in REMOTE_ACTIONS:
-                    self._log_message(
-                        f"    ⚠️ [{idx}/{cmd_count}] Unknown action: {action_name} (skipping)"
-                    )
-                    continue
-
-                # Get command template and substitute placeholders
-                command_template = REMOTE_ACTIONS[action_name]
-                command = command_template.format(
-                    timeout=timeout,
-                    path=cmd_config.path or ""
-                )
-                description = action_name
-
-                # Validate stop_compose has path
-                if action_name == "stop_compose" and not cmd_config.path:
-                    self._log_message(
-                        f"    ⚠️ [{idx}/{cmd_count}] stop_compose requires 'path' parameter (skipping)"
-                    )
-                    continue
-
-            # Handle custom command
-            elif cmd_config.command:
-                command = cmd_config.command
-                # Truncate long commands for display
-                if len(command) > 50:
-                    description = command[:47] + "..."
-                else:
-                    description = command
-
-            else:
-                self._log_message(
-                    f"    ⚠️ [{idx}/{cmd_count}] No action or command specified (skipping)"
-                )
-                continue
-
-            # Log what we're about to do
-            self._log_message(f"    ➡️ [{idx}/{cmd_count}] {description} (timeout: {timeout}s)")
-
-            if self.config.behavior.dry_run:
-                self._log_message(f"    🧪 [DRY-RUN] Would execute on {display_name}")
-                continue
-
-            # Execute the command
-            success, error_msg = self._run_remote_command(
-                server, command, timeout, description
-            )
-
-            if success:
-                self._log_message(f"    ✅ [{idx}/{cmd_count}] {description} completed")
-            else:
-                self._log_message(
-                    f"    ⚠️ [{idx}/{cmd_count}] {description} failed: {error_msg} (continuing)"
-                )
-
-        return True
-
-    def _shutdown_remote_server(self, server: RemoteServerConfig):
-        """Shutdown a single remote server via SSH.
-
-        Execution order:
-        1. Execute pre_shutdown_commands (if any) - best effort
-        2. Execute shutdown_command
-        """
-        display_name = server.name or server.host
-        has_pre_cmds = len(server.pre_shutdown_commands) > 0
-
-        self._log_message(f"🌐 Initiating remote shutdown: {display_name} ({server.host})...")
-
-        # Send notification for remote server shutdown start
-        self._send_notification(
-            f"🌐 **Remote Shutdown Starting:** {display_name}\n"
-            f"Host: {server.host}",
-            self.config.NOTIFY_INFO
-        )
-
-        # Execute pre-shutdown commands first
-        if has_pre_cmds:
-            self._execute_remote_pre_shutdown(server)
-
-        # Execute final shutdown command
-        self._log_message(f"  🔌 Sending shutdown command: {server.shutdown_command}")
-
-        if self.config.behavior.dry_run:
-            self._log_message(
-                f"  🧪 [DRY-RUN] Would send command '{server.shutdown_command}' to "
-                f"{server.user}@{server.host}"
-            )
-            return
-
-        success, error_msg = self._run_remote_command(
-            server,
-            server.shutdown_command,
-            server.command_timeout,
-            "shutdown"
-        )
-
-        if success:
-            self._log_message(f"  ✅ {display_name} shutdown command sent successfully")
-            self._send_notification(
-                f"✅ **Remote Shutdown Sent:** {display_name}\n"
-                f"Server is shutting down.",
-                self.config.NOTIFY_SUCCESS
-            )
-        else:
-            self._log_message(
-                f"  ❌ WARNING: Failed to execute shutdown command on {display_name}: {error_msg}"
-            )
-            self._send_notification(
-                f"❌ **Remote Shutdown Failed:** {display_name}\n"
-                f"Error: {error_msg}",
-                self.config.NOTIFY_FAILURE
-            )
 
     def _execute_shutdown_sequence(self):
         """Execute the controlled shutdown sequence."""
@@ -1296,173 +597,6 @@ class UPSGroupMonitor:
     # ==========================================================================
     # STATUS CHECKS
     # ==========================================================================
-
-    def _check_battery_anomaly(self, ups_data: Dict[str, str]):
-        """Detect abnormal battery charge changes while on line power.
-
-        Catches firmware recalibrations, battery aging events, or hardware
-        issues that cause sudden charge drops (e.g., 100% -> 60% in seconds)
-        while the UPS is on line power and not discharging.
-
-        Uses sustained-reading confirmation: an anomalous drop must persist
-        across 3 consecutive polls before firing.  This filters out transient
-        firmware jitter that some UPS units (notably APC, CyberPower, and
-        Ubiquiti UniFi UPS) exhibit after an OB -> OL transition, where the first
-        few readings may report a wildly incorrect charge that self-corrects
-        within a couple of seconds.
-        """
-        ups_status = ups_data.get('ups.status', '')
-        battery_charge_str = ups_data.get('battery.charge', '')
-
-        if not is_numeric(battery_charge_str):
-            return
-
-        current_charge = float(battery_charge_str)
-        current_time = time.time()
-
-        # Only track anomalies while on line power (OL/CHRG)
-        if "OB" in ups_status:
-            # On battery -- reset tracking, drops are expected
-            self.state.last_battery_charge = current_charge
-            self.state.last_battery_charge_time = current_time
-            self.state.pending_anomaly_charge = -1.0
-            self.state.pending_anomaly_count = 0
-            return
-
-        prev_charge = self.state.last_battery_charge
-        prev_time = self.state.last_battery_charge_time
-
-        # Update tracking
-        self.state.last_battery_charge = current_charge
-        self.state.last_battery_charge_time = current_time
-
-        # Skip if not yet initialized
-        if prev_charge < 0:
-            return
-
-        # Check for significant drop while online
-        drop = prev_charge - current_charge
-        elapsed = current_time - prev_time if prev_time > 0 else 0
-
-        # Threshold: >20% drop within 120 seconds while on line power
-        if drop > 20 and elapsed < 120:
-            # First detection -- record as pending, wait for confirmation
-            self.state.pending_anomaly_charge = current_charge
-            self.state.pending_anomaly_prev_charge = prev_charge
-            self.state.pending_anomaly_time = current_time
-            self.state.pending_anomaly_count = 1
-            return
-
-        # Check if a pending anomaly is being confirmed across polls
-        if self.state.pending_anomaly_charge >= 0:
-            # Charge recovered -- transient jitter, discard the anomaly
-            if current_charge > self.state.pending_anomaly_charge + 10:
-                self.state.pending_anomaly_charge = -1.0
-                self.state.pending_anomaly_count = 0
-                return
-
-            # Still low -- increment confirmation counter
-            self.state.pending_anomaly_count += 1
-
-            # Need 3 consecutive polls to confirm (filters firmware jitter)
-            if self.state.pending_anomaly_count < 3:
-                return
-
-            # Confirmed anomaly (sustained across 3 polls)
-            anomaly_prev = self.state.pending_anomaly_prev_charge
-            anomaly_drop = anomaly_prev - current_charge
-            anomaly_elapsed = current_time - self.state.pending_anomaly_time
-            self.state.pending_anomaly_charge = -1.0
-            self.state.pending_anomaly_count = 0
-
-            self._log_message(
-                f"⚠️ WARNING: Battery charge dropped from {anomaly_prev:.0f}% to "
-                f"{current_charge:.0f}% ({anomaly_drop:.0f}% drop) while on line power. "
-                f"Possible firmware recalibration, battery aging, or hardware issue."
-            )
-            self._send_notification(
-                f"⚠️ **Battery Anomaly Detected**\n"
-                f"Charge dropped from {anomaly_prev:.0f}% to {current_charge:.0f}% "
-                f"({anomaly_drop:.0f}% drop in {anomaly_elapsed:.0f}s) while on line power.\n"
-                f"Possible causes: firmware recalibration, battery aging, or hardware issue.",
-                self.config.NOTIFY_WARNING
-            )
-
-    def _check_voltage_issues(self, ups_status: str, input_voltage: str):
-        """Check for voltage quality issues."""
-        if "OL" not in ups_status:
-            if "OB" in ups_status or "FSD" in ups_status:
-                self.state.voltage_state = "NORMAL"
-            return
-
-        if not is_numeric(input_voltage):
-            return
-
-        voltage = float(input_voltage)
-
-        if voltage < self.state.voltage_warning_low:
-            if self.state.voltage_state != "LOW":
-                self._log_power_event(
-                    "BROWNOUT_DETECTED",
-                    f"Voltage is low: {voltage}V (Threshold: {self.state.voltage_warning_low}V)"
-                )
-                self.state.voltage_state = "LOW"
-        elif voltage > self.state.voltage_warning_high:
-            if self.state.voltage_state != "HIGH":
-                self._log_power_event(
-                    "OVER_VOLTAGE_DETECTED",
-                    f"Voltage is high: {voltage}V (Threshold: {self.state.voltage_warning_high}V)"
-                )
-                self.state.voltage_state = "HIGH"
-        elif self.state.voltage_state != "NORMAL":
-            self._log_power_event(
-                "VOLTAGE_NORMALIZED",
-                f"Voltage returned to normal: {voltage}V. Previous state: {self.state.voltage_state}"
-            )
-            self.state.voltage_state = "NORMAL"
-
-    def _check_avr_status(self, ups_status: str, input_voltage: str):
-        """Check for Automatic Voltage Regulation activity."""
-        voltage_str = f"{input_voltage}V" if is_numeric(input_voltage) else "N/A"
-
-        if "BOOST" in ups_status:
-            if self.state.avr_state != "BOOST":
-                self._log_power_event(
-                    "AVR_BOOST_ACTIVE",
-                    f"Input voltage low ({voltage_str}). UPS is boosting output."
-                )
-                self.state.avr_state = "BOOST"
-        elif "TRIM" in ups_status:
-            if self.state.avr_state != "TRIM":
-                self._log_power_event(
-                    "AVR_TRIM_ACTIVE",
-                    f"Input voltage high ({voltage_str}). UPS is trimming output."
-                )
-                self.state.avr_state = "TRIM"
-        elif self.state.avr_state != "INACTIVE":
-            self._log_power_event("AVR_INACTIVE", f"AVR is inactive. Input voltage: {voltage_str}.")
-            self.state.avr_state = "INACTIVE"
-
-    def _check_bypass_status(self, ups_status: str):
-        """Check for bypass mode."""
-        if "BYPASS" in ups_status:
-            if self.state.bypass_state != "ACTIVE":
-                self._log_power_event("BYPASS_MODE_ACTIVE", "UPS in bypass mode - no protection active!")
-                self.state.bypass_state = "ACTIVE"
-        elif self.state.bypass_state != "INACTIVE":
-            self._log_power_event("BYPASS_MODE_INACTIVE", "UPS left bypass mode.")
-            self.state.bypass_state = "INACTIVE"
-
-    def _check_overload_status(self, ups_status: str, ups_load: str):
-        """Check for overload condition."""
-        if "OVER" in ups_status:
-            if self.state.overload_state != "ACTIVE":
-                self._log_power_event("OVERLOAD_ACTIVE", f"UPS overload detected! Load: {ups_load}%")
-                self.state.overload_state = "ACTIVE"
-        elif self.state.overload_state != "INACTIVE":
-            reported_load = str(ups_load) if is_numeric(ups_load) else "N/A"
-            self._log_power_event("OVERLOAD_RESOLVED", f"UPS overload resolved. Load: {reported_load}%")
-            self.state.overload_state = "INACTIVE"
 
     def _handle_on_battery(self, ups_data: Dict[str, str]):
         """Handle the On Battery state."""
